OSIS-4646 move AjaxPermissionRequiredMixin in osis-role
